Struct aws_sdk_databasemigration::client::fluent_builders::CreateEndpoint

pub struct CreateEndpoint<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateEndpoint<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateEndpointOutput, SdkError<CreateEndpointError>> where
R::Policy: SmithyRetryPolicy<CreateEndpointInputOperationOutputAlias, CreateEndpointOutput, CreateEndpointError, CreateEndpointInputOperationRetryAlias>,

pub fn endpoint_identifier(self, inp: impl Into<String>) -> Self

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen, or contain two consecutive hyphens.

pub fn set_endpoint_identifier(self, input: Option<String>) -> Self

pub fn endpoint_type(self, inp: ReplicationEndpointTypeValue) -> Self

The type of endpoint. Valid values are source and target.

pub fn set_endpoint_type(
self,
input: Option<ReplicationEndpointTypeValue>
) -> Self

pub fn engine_name(self, inp: impl Into<String>) -> Self

The type of engine for the endpoint. Valid values, depending on the EndpointType value, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "docdb", "sqlserver", and "neptune".

pub fn set_engine_name(self, input: Option<String>) -> Self

pub fn username(self, inp: impl Into<String>) -> Self

The user name to be used to log in to the endpoint database.

pub fn set_username(self, input: Option<String>) -> Self

pub fn password(self, inp: impl Into<String>) -> Self

The password to be used to log in to the endpoint database.

pub fn set_password(self, input: Option<String>) -> Self

pub fn server_name(self, inp: impl Into<String>) -> Self

The name of the server where the endpoint database resides.

pub fn set_server_name(self, input: Option<String>) -> Self

pub fn port(self, inp: i32) -> Self

The port used by the endpoint database.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn database_name(self, inp: impl Into<String>) -> Self

The name of the endpoint database. For a MySQL source or target endpoint, do not specify DatabaseName.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn extra_connection_attributes(self, inp: impl Into<String>) -> Self

Additional attributes associated with the connection. Each attribute is specified as a name-value pair associated by an equal sign (=). Multiple attributes are separated by a semicolon (;) with no additional white space. For information on the attributes available for connecting your source or target endpoint, see Working with DMS Endpoints in the Database Migration Service User Guide.

pub fn set_extra_connection_attributes(self, input: Option<String>) -> Self

pub fn kms_key_id(self, inp: impl Into<String>) -> Self

An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

One or more tags to be assigned to the endpoint.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn certificate_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the certificate.

pub fn set_certificate_arn(self, input: Option<String>) -> Self

pub fn ssl_mode(self, inp: DmsSslModeValue) -> Self

The Secure Sockets Layer (SSL) mode to use for the SSL connection. The default is none

pub fn set_ssl_mode(self, input: Option<DmsSslModeValue>) -> Self

pub fn service_access_role_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the service access role that you want to use to create the endpoint. The role must allow the iam:PassRole action.

pub fn set_service_access_role_arn(self, input: Option<String>) -> Self

pub fn external_table_definition(self, inp: impl Into<String>) -> Self

The external table definition.

pub fn set_external_table_definition(self, input: Option<String>) -> Self

pub fn dynamo_db_settings(self, inp: DynamoDbSettings) -> Self

Settings in JSON format for the target Amazon DynamoDB endpoint. For information about other available settings, see Using Object Mapping to Migrate Data to DynamoDB in the Database Migration Service User Guide.

pub fn set_dynamo_db_settings(self, input: Option<DynamoDbSettings>) -> Self

pub fn s3_settings(self, inp: S3Settings) -> Self

Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see Extra Connection Attributes When Using Amazon S3 as a Target for DMS in the Database Migration Service User Guide.

pub fn set_s3_settings(self, input: Option<S3Settings>) -> Self

pub fn dms_transfer_settings(self, inp: DmsTransferSettings) -> Self

The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - The IAM role that has permission to access the Amazon S3 bucket. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string", }

pub fn set_dms_transfer_settings(
self,
input: Option<DmsTransferSettings>
) -> Self

pub fn mongo_db_settings(self, inp: MongoDbSettings) -> Self

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see Endpoint configuration settings when using MongoDB as a source for Database Migration Service in the Database Migration Service User Guide.

pub fn set_mongo_db_settings(self, input: Option<MongoDbSettings>) -> Self

pub fn kinesis_settings(self, inp: KinesisSettings) -> Self

Settings in JSON format for the target endpoint for Amazon Kinesis Data Streams. For more information about the available settings, see Using object mapping to migrate data to a Kinesis data stream in the Database Migration Service User Guide.

pub fn set_kinesis_settings(self, input: Option<KinesisSettings>) -> Self

pub fn kafka_settings(self, inp: KafkaSettings) -> Self

Settings in JSON format for the target Apache Kafka endpoint. For more information about the available settings, see Using object mapping to migrate data to a Kafka topic in the Database Migration Service User Guide.

pub fn set_kafka_settings(self, input: Option<KafkaSettings>) -> Self

pub fn elasticsearch_settings(self, inp: ElasticsearchSettings) -> Self

Settings in JSON format for the target Elasticsearch endpoint. For more information about the available settings, see Extra Connection Attributes When Using Elasticsearch as a Target for DMS in the Database Migration Service User Guide.

pub fn set_elasticsearch_settings(
self,
input: Option<ElasticsearchSettings>
) -> Self

pub fn neptune_settings(self, inp: NeptuneSettings) -> Self

Settings in JSON format for the target Amazon Neptune endpoint. For more information about the available settings, see Specifying graph-mapping rules using Gremlin and R2RML for Amazon Neptune as a target in the Database Migration Service User Guide.

pub fn set_neptune_settings(self, input: Option<NeptuneSettings>) -> Self

pub fn redshift_settings(self, inp: RedshiftSettings) -> Self

Provides information that defines an Amazon Redshift endpoint.

pub fn set_redshift_settings(self, input: Option<RedshiftSettings>) -> Self

pub fn postgre_sql_settings(self, inp: PostgreSqlSettings) -> Self

Settings in JSON format for the source and target PostgreSQL endpoint. For information about other available settings, see Extra connection attributes when using PostgreSQL as a source for DMS and Extra connection attributes when using PostgreSQL as a target for DMS in the Database Migration Service User Guide.

pub fn set_postgre_sql_settings(self, input: Option<PostgreSqlSettings>) -> Self

pub fn my_sql_settings(self, inp: MySqlSettings) -> Self

Settings in JSON format for the source and target MySQL endpoint. For information about other available settings, see Extra connection attributes when using MySQL as a source for DMS and Extra connection attributes when using a MySQL-compatible database as a target for DMS in the Database Migration Service User Guide.

pub fn set_my_sql_settings(self, input: Option<MySqlSettings>) -> Self

pub fn oracle_settings(self, inp: OracleSettings) -> Self

Settings in JSON format for the source and target Oracle endpoint. For information about other available settings, see Extra connection attributes when using Oracle as a source for DMS and Extra connection attributes when using Oracle as a target for DMS in the Database Migration Service User Guide.

pub fn set_oracle_settings(self, input: Option<OracleSettings>) -> Self

pub fn sybase_settings(self, inp: SybaseSettings) -> Self

Settings in JSON format for the source and target SAP ASE endpoint. For information about other available settings, see Extra connection attributes when using SAP ASE as a source for DMS and Extra connection attributes when using SAP ASE as a target for DMS in the Database Migration Service User Guide.

pub fn set_sybase_settings(self, input: Option<SybaseSettings>) -> Self

pub fn microsoft_sql_server_settings(
self,
inp: MicrosoftSqlServerSettings
) -> Self

Settings in JSON format for the source and target Microsoft SQL Server endpoint. For information about other available settings, see Extra connection attributes when using SQL Server as a source for DMS and Extra connection attributes when using SQL Server as a target for DMS in the Database Migration Service User Guide.

pub fn set_microsoft_sql_server_settings(
self,
input: Option<MicrosoftSqlServerSettings>
) -> Self

pub fn ibm_db2_settings(self, inp: IbmDb2Settings) -> Self

Settings in JSON format for the source IBM Db2 LUW endpoint. For information about other available settings, see Extra connection attributes when using Db2 LUW as a source for DMS in the Database Migration Service User Guide.

pub fn set_ibm_db2_settings(self, input: Option<IbmDb2Settings>) -> Self

pub fn resource_identifier(self, inp: impl Into<String>) -> Self

A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

pub fn set_resource_identifier(self, input: Option<String>) -> Self

pub fn doc_db_settings(self, inp: DocDbSettings) -> Self

Provides information that defines a DocumentDB endpoint.

pub fn set_doc_db_settings(self, input: Option<DocDbSettings>) -> Self

pub fn redis_settings(self, inp: RedisSettings) -> Self

Settings in JSON format for the target Redis endpoint.

pub fn set_redis_settings(self, input: Option<RedisSettings>) -> Self