Struct aws_sdk_databasemigration::client::fluent_builders::CreateReplicationTask

pub struct CreateReplicationTask<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateReplicationTask<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateReplicationTaskOutput, SdkError<CreateReplicationTaskError>> where
R::Policy: SmithyRetryPolicy<CreateReplicationTaskInputOperationOutputAlias, CreateReplicationTaskOutput, CreateReplicationTaskError, CreateReplicationTaskInputOperationRetryAlias>,

pub fn replication_task_identifier(self, inp: impl Into<String>) -> Self

An identifier for the replication task.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

pub fn set_replication_task_identifier(self, input: Option<String>) -> Self

pub fn source_endpoint_arn(self, inp: impl Into<String>) -> Self

An Amazon Resource Name (ARN) that uniquely identifies the source endpoint.

pub fn set_source_endpoint_arn(self, input: Option<String>) -> Self

pub fn target_endpoint_arn(self, inp: impl Into<String>) -> Self

An Amazon Resource Name (ARN) that uniquely identifies the target endpoint.

pub fn set_target_endpoint_arn(self, input: Option<String>) -> Self

pub fn replication_instance_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of a replication instance.

pub fn set_replication_instance_arn(self, input: Option<String>) -> Self

pub fn migration_type(self, inp: MigrationTypeValue) -> Self

The migration type. Valid values: full-load | cdc | full-load-and-cdc

pub fn set_migration_type(self, input: Option<MigrationTypeValue>) -> Self

pub fn table_mappings(self, inp: impl Into<String>) -> Self

The table mappings for the task, in JSON format. For more information, see Using Table Mapping to Specify Task Settings in the Database Migration Service User Guide.

pub fn set_table_mappings(self, input: Option<String>) -> Self

pub fn replication_task_settings(self, inp: impl Into<String>) -> Self

Overall settings for the task, in JSON format. For more information, see Specifying Task Settings for Database Migration Service Tasks in the Database Migration Service User Guide.

pub fn set_replication_task_settings(self, input: Option<String>) -> Self

pub fn cdc_start_time(self, inp: Instant) -> Self

Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”

pub fn set_cdc_start_time(self, input: Option<Instant>) -> Self

pub fn cdc_start_position(self, inp: impl Into<String>) -> Self

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1#27#mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:1876#0#0#*#0#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a logical replication slot should already be created and associated with the source endpoint. You can verify this by setting the slotName extra connection attribute to the name of this logical replication slot. For more information, see Extra Connection Attributes When Using PostgreSQL as a Source for DMS.

pub fn set_cdc_start_position(self, input: Option<String>) -> Self

pub fn cdc_stop_position(self, inp: impl Into<String>) -> Self

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

pub fn set_cdc_stop_position(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

One or more tags to be assigned to the replication task.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn task_data(self, inp: impl Into<String>) -> Self

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

pub fn set_task_data(self, input: Option<String>) -> Self

pub fn resource_identifier(self, inp: impl Into<String>) -> Self

A friendly name for the resource identifier at the end of the EndpointArn response parameter that is returned in the created Endpoint object. The value for this parameter can have up to 31 characters. It can contain only ASCII letters, digits, and hyphen ('-'). Also, it can't end with a hyphen or contain two consecutive hyphens, and can only begin with a letter, such as Example-App-ARN1. For example, this value might result in the EndpointArn value arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1. If you don't specify a ResourceIdentifier value, DMS generates a default identifier value for the end of EndpointArn.

pub fn set_resource_identifier(self, input: Option<String>) -> Self