Struct aws_sdk_databasemigration::client::fluent_builders::ModifyReplicationInstance

pub struct ModifyReplicationInstance<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> ModifyReplicationInstance<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<ModifyReplicationInstanceOutput, SdkError<ModifyReplicationInstanceError>> where
R::Policy: SmithyRetryPolicy<ModifyReplicationInstanceInputOperationOutputAlias, ModifyReplicationInstanceOutput, ModifyReplicationInstanceError, ModifyReplicationInstanceInputOperationRetryAlias>,

pub fn replication_instance_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the replication instance.

pub fn set_replication_instance_arn(self, input: Option<String>) -> Self

pub fn allocated_storage(self, inp: i32) -> Self

The amount of storage (in gigabytes) to be allocated for the replication instance.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn apply_immediately(self, inp: bool) -> Self

Indicates whether the changes should be applied immediately or during the next maintenance window.

pub fn set_apply_immediately(self, input: Option<bool>) -> Self

pub fn replication_instance_class(self, inp: impl Into<String>) -> Self

The compute and memory capacity of the replication instance as defined for the specified replication instance class. For example to specify the instance class dms.c4.large, set this parameter to "dms.c4.large".

For more information on the settings and capacities for the available replication instance classes, see Selecting the right DMS replication instance for your migration.

pub fn set_replication_instance_class(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to VpcSecurityGroupIds.

To override the contents of this collection use set_vpc_security_group_ids.

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn preferred_maintenance_window(self, inp: impl Into<String>) -> Self

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter does not result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn multi_az(self, inp: bool) -> Self

Specifies whether the replication instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine_version(self, inp: impl Into<String>) -> Self

The engine version number of the replication instance.

When modifying a major engine version of an instance, also set AllowMajorVersionUpgrade to true.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn allow_major_version_upgrade(self, inp: bool) -> Self

Indicates that major version upgrades are allowed. Changing this parameter does not result in an outage, and the change is asynchronously applied as soon as possible.

This parameter must be set to true when specifying a value for the EngineVersion parameter that is a different major version than the replication instance's current version.

pub fn set_allow_major_version_upgrade(self, input: Option<bool>) -> Self

pub fn auto_minor_version_upgrade(self, inp: bool) -> Self

A value that indicates that minor version upgrades are applied automatically to the replication instance during the maintenance window. Changing this parameter doesn't result in an outage, except in the case described following. The change is asynchronously applied as soon as possible.

An outage does result if these factors apply:

This parameter is set to true during the maintenance window.

A newer minor version is available.

DMS has enabled automatic patching for the given engine version.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn replication_instance_identifier(self, inp: impl Into<String>) -> Self

The replication instance identifier. This parameter is stored as a lowercase string.

pub fn set_replication_instance_identifier(self, input: Option<String>) -> Self