Module aws_sdk_databasemigration::error

Modules

access_denied_fault

See AccessDeniedFault

insufficient_resource_capacity_fault

See InsufficientResourceCapacityFault

invalid_certificate_fault

See InvalidCertificateFault

invalid_resource_state_fault

See InvalidResourceStateFault

invalid_subnet

See InvalidSubnet

kms_access_denied_fault

See KmsAccessDeniedFault

kms_disabled_fault

See KmsDisabledFault

kms_fault

See KmsFault

kms_invalid_state_fault

See KmsInvalidStateFault

kms_key_not_accessible_fault

See KmsKeyNotAccessibleFault

kms_not_found_fault

See KmsNotFoundFault

kms_throttling_fault

See KmsThrottlingFault

replication_subnet_group_does_not_cover_enough_a_zs

See ReplicationSubnetGroupDoesNotCoverEnoughAZs

resource_already_exists_fault

See ResourceAlreadyExistsFault

resource_not_found_fault

See ResourceNotFoundFault

resource_quota_exceeded_fault

See ResourceQuotaExceededFault

s3_access_denied_fault

See S3AccessDeniedFault

s3_resource_not_found_fault

See S3ResourceNotFoundFault

sns_invalid_topic_fault

See SnsInvalidTopicFault

sns_no_authorization_fault

See SnsNoAuthorizationFault

storage_quota_exceeded_fault

See StorageQuotaExceededFault

subnet_already_in_use

See SubnetAlreadyInUse

upgrade_dependency_failure_fault

See UpgradeDependencyFailureFault

Structs

AccessDeniedFault

DMS was denied access to the endpoint. Check that the

AddTagsToResourceError

ApplyPendingMaintenanceActionError

CancelReplicationTaskAssessmentRunError

CreateEndpointError

CreateEventSubscriptionError

CreateReplicationInstanceError

CreateReplicationSubnetGroupError

CreateReplicationTaskError

DeleteCertificateError

DeleteConnectionError

DeleteEndpointError

DeleteEventSubscriptionError

DeleteReplicationInstanceError

DeleteReplicationSubnetGroupError

DeleteReplicationTaskAssessmentRunError

DeleteReplicationTaskError

DescribeAccountAttributesError

DescribeApplicableIndividualAssessmentsError

DescribeCertificatesError

DescribeConnectionsError

DescribeEndpointSettingsError

DescribeEndpointTypesError

DescribeEndpointsError

DescribeEventCategoriesError

DescribeEventSubscriptionsError

DescribeEventsError

DescribeOrderableReplicationInstancesError

DescribePendingMaintenanceActionsError

DescribeRefreshSchemasStatusError

DescribeReplicationInstanceTaskLogsError

DescribeReplicationInstancesError

DescribeReplicationSubnetGroupsError

DescribeReplicationTaskAssessmentResultsError

DescribeReplicationTaskAssessmentRunsError

DescribeReplicationTaskIndividualAssessmentsError

DescribeReplicationTasksError

DescribeSchemasError

DescribeTableStatisticsError

ImportCertificateError

InsufficientResourceCapacityFault

There are not enough resources allocated to the database migration.

InvalidCertificateFault

The certificate was not valid.

InvalidResourceStateFault

The resource is in a state that prevents it from being used for database migration.

InvalidSubnet

The subnet provided is invalid.

KmsAccessDeniedFault

The ciphertext references a key that doesn't exist or that the DMS account doesn't have access to.

KmsDisabledFault

The specified master key (CMK) isn't enabled.

KmsFault

An Key Management Service (KMS) error is preventing access to KMS.

KmsInvalidStateFault

The state of the specified KMS resource isn't valid for this request.

KmsKeyNotAccessibleFault

DMS cannot access the KMS key.

KmsNotFoundFault

The specified KMS entity or resource can't be found.

KmsThrottlingFault

This request triggered KMS request throttling.

ListTagsForResourceError

ModifyEndpointError

ModifyEventSubscriptionError

ModifyReplicationInstanceError

ModifyReplicationSubnetGroupError

ModifyReplicationTaskError

MoveReplicationTaskError

RebootReplicationInstanceError

RefreshSchemasError

ReloadTablesError

RemoveTagsFromResourceError

ReplicationSubnetGroupDoesNotCoverEnoughAZs

The replication subnet group does not cover enough Availability Zones (AZs). Edit the replication subnet group and add more AZs.

ResourceAlreadyExistsFault

The resource you are attempting to create already exists.

ResourceNotFoundFault

The resource could not be found.

ResourceQuotaExceededFault

The quota for this resource quota has been exceeded.

S3AccessDeniedFault

Insufficient privileges are preventing access to an Amazon S3 object.

S3ResourceNotFoundFault

A specified Amazon S3 bucket, bucket folder, or other object can't be

SnsInvalidTopicFault

The SNS topic is invalid.

SnsNoAuthorizationFault

You are not authorized for the SNS subscription.

StartReplicationTaskAssessmentError

StartReplicationTaskAssessmentRunError

StartReplicationTaskError

StopReplicationTaskError

StorageQuotaExceededFault

The storage quota has been exceeded.

SubnetAlreadyInUse

The specified subnet is already in use.

TestConnectionError

UpgradeDependencyFailureFault

An upgrade dependency is preventing the database migration.

Enums

AddTagsToResourceErrorKind

ApplyPendingMaintenanceActionErrorKind

CancelReplicationTaskAssessmentRunErrorKind

CreateEndpointErrorKind

CreateEventSubscriptionErrorKind

CreateReplicationInstanceErrorKind

CreateReplicationSubnetGroupErrorKind

CreateReplicationTaskErrorKind

DeleteCertificateErrorKind

DeleteConnectionErrorKind

DeleteEndpointErrorKind

DeleteEventSubscriptionErrorKind

DeleteReplicationInstanceErrorKind

DeleteReplicationSubnetGroupErrorKind

DeleteReplicationTaskAssessmentRunErrorKind

DeleteReplicationTaskErrorKind

DescribeAccountAttributesErrorKind

DescribeApplicableIndividualAssessmentsErrorKind

DescribeCertificatesErrorKind

DescribeConnectionsErrorKind

DescribeEndpointSettingsErrorKind

DescribeEndpointTypesErrorKind

DescribeEndpointsErrorKind

DescribeEventCategoriesErrorKind

DescribeEventSubscriptionsErrorKind

DescribeEventsErrorKind

DescribeOrderableReplicationInstancesErrorKind

DescribePendingMaintenanceActionsErrorKind

DescribeRefreshSchemasStatusErrorKind

DescribeReplicationInstanceTaskLogsErrorKind

DescribeReplicationInstancesErrorKind

DescribeReplicationSubnetGroupsErrorKind

DescribeReplicationTaskAssessmentResultsErrorKind

DescribeReplicationTaskAssessmentRunsErrorKind

DescribeReplicationTaskIndividualAssessmentsErrorKind

DescribeReplicationTasksErrorKind

DescribeSchemasErrorKind

DescribeTableStatisticsErrorKind

ImportCertificateErrorKind

ListTagsForResourceErrorKind

ModifyEndpointErrorKind

ModifyEventSubscriptionErrorKind

ModifyReplicationInstanceErrorKind

ModifyReplicationSubnetGroupErrorKind

ModifyReplicationTaskErrorKind

MoveReplicationTaskErrorKind

RebootReplicationInstanceErrorKind

RefreshSchemasErrorKind

ReloadTablesErrorKind

RemoveTagsFromResourceErrorKind

StartReplicationTaskAssessmentErrorKind

StartReplicationTaskAssessmentRunErrorKind

StartReplicationTaskErrorKind

StopReplicationTaskErrorKind

TestConnectionErrorKind