Struct aws_sdk_databasemigration::input::modify_endpoint_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ModifyEndpointInput

Implementations

impl Builder

pub fn endpoint_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

pub fn set_endpoint_arn(self, input: Option<String>) -> Self

pub fn endpoint_identifier(self, input: impl Into<String>) -> Self

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

pub fn set_endpoint_identifier(self, input: Option<String>) -> Self

pub fn endpoint_type(self, input: ReplicationEndpointTypeValue) -> Self

The type of endpoint. Valid values are source and target.

pub fn set_endpoint_type(
self,
input: Option<ReplicationEndpointTypeValue>
) -> Self

pub fn engine_name(self, input: impl Into<String>) -> Self

The type of engine for the endpoint. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

pub fn set_engine_name(self, input: Option<String>) -> Self

pub fn username(self, input: impl Into<String>) -> Self

The user name to be used to login to the endpoint database.

pub fn set_username(self, input: Option<String>) -> Self

pub fn password(self, input: impl Into<String>) -> Self

The password to be used to login to the endpoint database.

pub fn set_password(self, input: Option<String>) -> Self

pub fn server_name(self, input: impl Into<String>) -> Self

The name of the server where the endpoint database resides.

pub fn set_server_name(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port used by the endpoint database.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

The name of the endpoint database. For a MySQL source or target endpoint, do not specify DatabaseName.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn extra_connection_attributes(self, input: impl Into<String>) -> Self

Additional attributes associated with the connection. To reset this parameter, pass the empty string ("") as an argument.

pub fn set_extra_connection_attributes(self, input: Option<String>) -> Self

pub fn certificate_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the certificate used for SSL connection.

pub fn set_certificate_arn(self, input: Option<String>) -> Self

pub fn ssl_mode(self, input: DmsSslModeValue) -> Self

The SSL mode used to connect to the endpoint. The default value is none.

pub fn set_ssl_mode(self, input: Option<DmsSslModeValue>) -> Self

pub fn service_access_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the IAM role you want to use to modify the endpoint. The role must allow the iam:PassRole action.

pub fn set_service_access_role_arn(self, input: Option<String>) -> Self

pub fn external_table_definition(self, input: impl Into<String>) -> Self

The external table definition.

pub fn set_external_table_definition(self, input: Option<String>) -> Self

pub fn dynamo_db_settings(self, input: DynamoDbSettings) -> Self

Settings in JSON format for the target Amazon DynamoDB endpoint. For information about other available settings, see Using Object Mapping to Migrate Data to DynamoDB in the Database Migration Service User Guide.

pub fn set_dynamo_db_settings(self, input: Option<DynamoDbSettings>) -> Self

pub fn s3_settings(self, input: S3Settings) -> Self

Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see Extra Connection Attributes When Using Amazon S3 as a Target for DMS in the Database Migration Service User Guide.

pub fn set_s3_settings(self, input: Option<S3Settings>) -> Self

pub fn dms_transfer_settings(self, input: DmsTransferSettings) -> Self

The settings in JSON format for the DMS transfer type of source endpoint.

Attributes include the following:

serviceAccessRoleArn - The Identity and Access Management (IAM) role that has permission to access the Amazon S3 bucket. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string ,BucketName=string

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

pub fn set_dms_transfer_settings(
self,
input: Option<DmsTransferSettings>
) -> Self

pub fn mongo_db_settings(self, input: MongoDbSettings) -> Self

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the configuration properties section in Endpoint configuration settings when using MongoDB as a source for Database Migration Service in the Database Migration Service User Guide.

pub fn set_mongo_db_settings(self, input: Option<MongoDbSettings>) -> Self

pub fn kinesis_settings(self, input: KinesisSettings) -> Self

Settings in JSON format for the target endpoint for Amazon Kinesis Data Streams. For more information about the available settings, see Using object mapping to migrate data to a Kinesis data stream in the Database Migration Service User Guide.

pub fn set_kinesis_settings(self, input: Option<KinesisSettings>) -> Self

pub fn kafka_settings(self, input: KafkaSettings) -> Self

Settings in JSON format for the target Apache Kafka endpoint. For more information about the available settings, see Using object mapping to migrate data to a Kafka topic in the Database Migration Service User Guide.

pub fn set_kafka_settings(self, input: Option<KafkaSettings>) -> Self

pub fn elasticsearch_settings(self, input: ElasticsearchSettings) -> Self

Settings in JSON format for the target Elasticsearch endpoint. For more information about the available settings, see Extra Connection Attributes When Using Elasticsearch as a Target for DMS in the Database Migration Service User Guide.

pub fn set_elasticsearch_settings(
self,
input: Option<ElasticsearchSettings>
) -> Self

pub fn neptune_settings(self, input: NeptuneSettings) -> Self

Settings in JSON format for the target Amazon Neptune endpoint. For more information about the available settings, see Specifying graph-mapping rules using Gremlin and R2RML for Amazon Neptune as a target in the Database Migration Service User Guide.

pub fn set_neptune_settings(self, input: Option<NeptuneSettings>) -> Self

pub fn redshift_settings(self, input: RedshiftSettings) -> Self

Provides information that defines an Amazon Redshift endpoint.

pub fn set_redshift_settings(self, input: Option<RedshiftSettings>) -> Self

pub fn postgre_sql_settings(self, input: PostgreSqlSettings) -> Self

Settings in JSON format for the source and target PostgreSQL endpoint. For information about other available settings, see Extra connection attributes when using PostgreSQL as a source for DMS and Extra connection attributes when using PostgreSQL as a target for DMS in the Database Migration Service User Guide.

pub fn set_postgre_sql_settings(self, input: Option<PostgreSqlSettings>) -> Self

pub fn my_sql_settings(self, input: MySqlSettings) -> Self

Settings in JSON format for the source and target MySQL endpoint. For information about other available settings, see Extra connection attributes when using MySQL as a source for DMS and Extra connection attributes when using a MySQL-compatible database as a target for DMS in the Database Migration Service User Guide.

pub fn set_my_sql_settings(self, input: Option<MySqlSettings>) -> Self

pub fn oracle_settings(self, input: OracleSettings) -> Self

Settings in JSON format for the source and target Oracle endpoint. For information about other available settings, see Extra connection attributes when using Oracle as a source for DMS and Extra connection attributes when using Oracle as a target for DMS in the Database Migration Service User Guide.

pub fn set_oracle_settings(self, input: Option<OracleSettings>) -> Self

pub fn sybase_settings(self, input: SybaseSettings) -> Self

Settings in JSON format for the source and target SAP ASE endpoint. For information about other available settings, see Extra connection attributes when using SAP ASE as a source for DMS and Extra connection attributes when using SAP ASE as a target for DMS in the Database Migration Service User Guide.

pub fn set_sybase_settings(self, input: Option<SybaseSettings>) -> Self

pub fn microsoft_sql_server_settings(
self,
input: MicrosoftSqlServerSettings
) -> Self

Settings in JSON format for the source and target Microsoft SQL Server endpoint. For information about other available settings, see Extra connection attributes when using SQL Server as a source for DMS and Extra connection attributes when using SQL Server as a target for DMS in the Database Migration Service User Guide.

pub fn set_microsoft_sql_server_settings(
self,
input: Option<MicrosoftSqlServerSettings>
) -> Self

pub fn ibm_db2_settings(self, input: IbmDb2Settings) -> Self

Settings in JSON format for the source IBM Db2 LUW endpoint. For information about other available settings, see Extra connection attributes when using Db2 LUW as a source for DMS in the Database Migration Service User Guide.

pub fn set_ibm_db2_settings(self, input: Option<IbmDb2Settings>) -> Self

pub fn doc_db_settings(self, input: DocDbSettings) -> Self

Settings in JSON format for the source DocumentDB endpoint. For more information about the available settings, see the configuration properties section in Using DocumentDB as a Target for Database Migration Service in the Database Migration Service User Guide.

pub fn set_doc_db_settings(self, input: Option<DocDbSettings>) -> Self

pub fn redis_settings(self, input: RedisSettings) -> Self

Settings in JSON format for the Redis target endpoint.

pub fn set_redis_settings(self, input: Option<RedisSettings>) -> Self

pub fn exact_settings(self, input: bool) -> Self

If this attribute is Y, the current call to ModifyEndpoint replaces all existing endpoint settings with the exact settings that you specify in this call. If this attribute is N, the current call to ModifyEndpoint does two things:

It replaces any endpoint settings that already exist with new values, for settings with the same names.

It creates new endpoint settings that you specify in the call, for settings with different names.

For example, if you call create-endpoint ... --endpoint-settings '{"a":1}' ..., the endpoint has the following endpoint settings: '{"a":1}'. If you then call modify-endpoint ... --endpoint-settings '{"b":2}' ... for the same endpoint, the endpoint has the following settings: '{"a":1,"b":2}'.

However, suppose that you follow this with a call to modify-endpoint ... --endpoint-settings '{"b":2}' --exact-settings ... for that same endpoint again. Then the endpoint has the following settings: '{"b":2}'. All existing settings are replaced with the exact settings that you specify.

pub fn set_exact_settings(self, input: Option<bool>) -> Self

pub fn build(self) -> Result<ModifyEndpointInput, BuildError>

Consumes the builder and constructs a ModifyEndpointInput