Module aws_sdk_databasemigration::model

Modules

account_quota

See AccountQuota

availability_zone

See AvailabilityZone

certificate

See Certificate

connection

See Connection

dms_transfer_settings

See DmsTransferSettings

doc_db_settings

See DocDbSettings

dynamo_db_settings

See DynamoDbSettings

elasticsearch_settings

See ElasticsearchSettings

endpoint

See Endpoint

endpoint_setting

See EndpointSetting

event

See Event

event_category_group

See EventCategoryGroup

event_subscription

See EventSubscription

filter

See Filter

ibm_db2_settings

See IbmDb2Settings

kafka_settings

See KafkaSettings

kinesis_settings

See KinesisSettings

microsoft_sql_server_settings

See MicrosoftSqlServerSettings

mongo_db_settings

See MongoDbSettings

my_sql_settings

See MySqlSettings

neptune_settings

See NeptuneSettings

oracle_settings

See OracleSettings

orderable_replication_instance

See OrderableReplicationInstance

pending_maintenance_action

See PendingMaintenanceAction

postgre_sql_settings

See PostgreSqlSettings

redis_settings

See RedisSettings

redshift_settings

See RedshiftSettings

refresh_schemas_status

See RefreshSchemasStatus

replication_instance

See ReplicationInstance

replication_instance_task_log

See ReplicationInstanceTaskLog

replication_pending_modified_values

See ReplicationPendingModifiedValues

replication_subnet_group

See ReplicationSubnetGroup

replication_task

See ReplicationTask

replication_task_assessment_result

See ReplicationTaskAssessmentResult

replication_task_assessment_run

See ReplicationTaskAssessmentRun

replication_task_assessment_run_progress

See ReplicationTaskAssessmentRunProgress

replication_task_individual_assessment

See ReplicationTaskIndividualAssessment

replication_task_stats

See ReplicationTaskStats

resource_pending_maintenance_actions

See ResourcePendingMaintenanceActions

s3_settings

See S3Settings

subnet

See Subnet

supported_endpoint_type

See SupportedEndpointType

sybase_settings

See SybaseSettings

table_statistics

See TableStatistics

table_to_reload

See TableToReload

tag

See Tag

vpc_security_group_membership

See VpcSecurityGroupMembership

Structs

AccountQuota

Describes a quota for an Amazon Web Services account, for example the number of replication instances

AvailabilityZone

The name of an Availability Zone for use during database migration.

Certificate

The SSL certificate that can be used to encrypt connections between the endpoints and

Connection

Status of the connection between an endpoint and a replication instance, including

DmsTransferSettings

The settings in JSON format for the DMS Transfer type source endpoint.

DocDbSettings

Provides information that defines a DocumentDB endpoint.

DynamoDbSettings

Provides the Amazon Resource Name (ARN) of the Identity and Access Management (IAM)

ElasticsearchSettings

Provides information that defines an Elasticsearch endpoint.

Endpoint

Describes an endpoint of a database instance in response to operations such as the

EndpointSetting

Endpoint settings.

Event

Describes an identifiable significant activity that affects a replication instance or

EventCategoryGroup

Lists categories of events subscribed to, and generated by, the applicable DMS

EventSubscription

Describes an event notification subscription created by the CreateEventSubscription

Filter

Identifies the name and value of a filter object. This filter is used to limit the

IbmDb2Settings

Provides information that defines an IBM Db2 LUW endpoint.

KafkaSettings

Provides information that describes an Apache Kafka endpoint. This

KinesisSettings

Provides information that describes an Amazon Kinesis Data Stream endpoint. This

MicrosoftSqlServerSettings

Provides information that defines a Microsoft SQL Server endpoint.

MongoDbSettings

Provides information that defines a MongoDB endpoint.

MySqlSettings

Provides information that defines a MySQL endpoint.

NeptuneSettings

Provides information that defines an Amazon Neptune endpoint.

OracleSettings

Provides information that defines an Oracle endpoint.

OrderableReplicationInstance

In response to the DescribeOrderableReplicationInstances operation, this

PendingMaintenanceAction

Describes a maintenance action pending for an DMS resource, including when and how

PostgreSqlSettings

Provides information that defines a PostgreSQL endpoint.

RedisSettings

Provides information that defines a Redis target endpoint.

RedshiftSettings

Provides information that defines an Amazon Redshift endpoint.

RefreshSchemasStatus

Provides information that describes status of a schema at an endpoint specified by the

ReplicationInstance

Provides information that defines a replication instance.

ReplicationInstanceTaskLog

Contains metadata for a replication instance task log.

ReplicationPendingModifiedValues

Provides information about the values of pending modifications to a replication

ReplicationSubnetGroup

Describes a subnet group in response to a request by the

ReplicationTask

Provides information that describes a replication task created by the

ReplicationTaskAssessmentResult

The task assessment report in JSON format.

ReplicationTaskAssessmentRun

Provides information that describes a premigration assessment run that you have started

ReplicationTaskAssessmentRunProgress

The progress values reported by the AssessmentProgress response

ReplicationTaskIndividualAssessment

Provides information that describes an individual assessment from a premigration

ReplicationTaskStats

In response to a request by the DescribeReplicationTasks operation, this object provides

ResourcePendingMaintenanceActions

Identifies an DMS resource and any pending actions for it.

S3Settings

Settings for exporting data to Amazon S3.

Subnet

In response to a request by the DescribeReplicationSubnetGroups operation,

SupportedEndpointType

Provides information about types of supported endpoints in response to a request by the

SybaseSettings

Provides information that defines a SAP ASE endpoint.

TableStatistics

Provides a collection of table statistics in response to a request by the

TableToReload

Provides the name of the schema and table to be reloaded.

Tag

A user-defined key-value pair that describes metadata added to an DMS resource and

VpcSecurityGroupMembership

Describes the status of a security group associated with the virtual private cloud (VPC)

Enums

AuthMechanismValue

AuthTypeValue

CannedAclForObjectsValue

CharLengthSemantics

CompressionTypeValue

DataFormatValue

DatePartitionDelimiterValue

DatePartitionSequenceValue

DmsSslModeValue

EncodingTypeValue

EncryptionModeValue

EndpointSettingTypeValue

KafkaSecurityProtocol

MessageFormatValue

MigrationTypeValue

NestingLevelValue

ParquetVersionValue

PluginNameValue

RedisAuthTypeValue

RefreshSchemasStatusTypeValue

ReleaseStatusValues

ReloadOptionValue

ReplicationEndpointTypeValue

SafeguardPolicy

SourceType

SslSecurityProtocolValue

StartReplicationTaskTypeValue

TargetDbType