Struct aws_sdk_databasemigration::model::endpoint::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Endpoint

Implementations

impl Builder

pub fn endpoint_identifier(self, input: impl Into<String>) -> Self

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

pub fn set_endpoint_identifier(self, input: Option<String>) -> Self

pub fn endpoint_type(self, input: ReplicationEndpointTypeValue) -> Self

The type of endpoint. Valid values are source and target.

pub fn set_endpoint_type(
self,
input: Option<ReplicationEndpointTypeValue>
) -> Self

pub fn engine_name(self, input: impl Into<String>) -> Self

The database engine name. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

pub fn set_engine_name(self, input: Option<String>) -> Self

pub fn engine_display_name(self, input: impl Into<String>) -> Self

The expanded name for the engine name. For example, if the EngineName parameter is "aurora," this value would be "Amazon Aurora MySQL."

pub fn set_engine_display_name(self, input: Option<String>) -> Self

pub fn username(self, input: impl Into<String>) -> Self

The user name used to connect to the endpoint.

pub fn set_username(self, input: Option<String>) -> Self

pub fn server_name(self, input: impl Into<String>) -> Self

The name of the server at the endpoint.

pub fn set_server_name(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port value used to access the endpoint.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

The name of the database at the endpoint.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn extra_connection_attributes(self, input: impl Into<String>) -> Self

Additional connection attributes used to connect to the endpoint.

pub fn set_extra_connection_attributes(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The status of the endpoint.

pub fn set_status(self, input: Option<String>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn endpoint_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

pub fn set_endpoint_arn(self, input: Option<String>) -> Self

pub fn certificate_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) used for SSL connection to the endpoint.

pub fn set_certificate_arn(self, input: Option<String>) -> Self

pub fn ssl_mode(self, input: DmsSslModeValue) -> Self

The SSL mode used to connect to the endpoint. The default value is none.

pub fn set_ssl_mode(self, input: Option<DmsSslModeValue>) -> Self

pub fn service_access_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) used by the service to access the IAM role. The role must allow the iam:PassRole action.

pub fn set_service_access_role_arn(self, input: Option<String>) -> Self

pub fn external_table_definition(self, input: impl Into<String>) -> Self

The external table definition.

pub fn set_external_table_definition(self, input: Option<String>) -> Self

pub fn external_id(self, input: impl Into<String>) -> Self

Value returned by a call to CreateEndpoint that can be used for cross-account validation. Use it on a subsequent call to CreateEndpoint to create the endpoint with a cross-account.

pub fn set_external_id(self, input: Option<String>) -> Self

pub fn dynamo_db_settings(self, input: DynamoDbSettings) -> Self

The settings for the DynamoDB target endpoint. For more information, see the DynamoDBSettings structure.

pub fn set_dynamo_db_settings(self, input: Option<DynamoDbSettings>) -> Self

pub fn s3_settings(self, input: S3Settings) -> Self

The settings for the S3 target endpoint. For more information, see the S3Settings structure.

pub fn set_s3_settings(self, input: Option<S3Settings>) -> Self

pub fn dms_transfer_settings(self, input: DmsTransferSettings) -> Self

The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - The IAM role that has permission to access the Amazon S3 bucket. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string,

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

pub fn set_dms_transfer_settings(
self,
input: Option<DmsTransferSettings>
) -> Self

pub fn mongo_db_settings(self, input: MongoDbSettings) -> Self

The settings for the MongoDB source endpoint. For more information, see the MongoDbSettings structure.

pub fn set_mongo_db_settings(self, input: Option<MongoDbSettings>) -> Self

pub fn kinesis_settings(self, input: KinesisSettings) -> Self

The settings for the Amazon Kinesis target endpoint. For more information, see the KinesisSettings structure.

pub fn set_kinesis_settings(self, input: Option<KinesisSettings>) -> Self

pub fn kafka_settings(self, input: KafkaSettings) -> Self

The settings for the Apache Kafka target endpoint. For more information, see the KafkaSettings structure.

pub fn set_kafka_settings(self, input: Option<KafkaSettings>) -> Self

pub fn elasticsearch_settings(self, input: ElasticsearchSettings) -> Self

The settings for the Elasticsearch source endpoint. For more information, see the ElasticsearchSettings structure.

pub fn set_elasticsearch_settings(
self,
input: Option<ElasticsearchSettings>
) -> Self

pub fn neptune_settings(self, input: NeptuneSettings) -> Self

The settings for the Amazon Neptune target endpoint. For more information, see the NeptuneSettings structure.

pub fn set_neptune_settings(self, input: Option<NeptuneSettings>) -> Self

pub fn redshift_settings(self, input: RedshiftSettings) -> Self

Settings for the Amazon Redshift endpoint.

pub fn set_redshift_settings(self, input: Option<RedshiftSettings>) -> Self

pub fn postgre_sql_settings(self, input: PostgreSqlSettings) -> Self

The settings for the PostgreSQL source and target endpoint. For more information, see the PostgreSQLSettings structure.

pub fn set_postgre_sql_settings(self, input: Option<PostgreSqlSettings>) -> Self

pub fn my_sql_settings(self, input: MySqlSettings) -> Self

The settings for the MySQL source and target endpoint. For more information, see the MySQLSettings structure.

pub fn set_my_sql_settings(self, input: Option<MySqlSettings>) -> Self

pub fn oracle_settings(self, input: OracleSettings) -> Self

The settings for the Oracle source and target endpoint. For more information, see the OracleSettings structure.

pub fn set_oracle_settings(self, input: Option<OracleSettings>) -> Self

pub fn sybase_settings(self, input: SybaseSettings) -> Self

The settings for the SAP ASE source and target endpoint. For more information, see the SybaseSettings structure.

pub fn set_sybase_settings(self, input: Option<SybaseSettings>) -> Self

pub fn microsoft_sql_server_settings(
self,
input: MicrosoftSqlServerSettings
) -> Self

The settings for the Microsoft SQL Server source and target endpoint. For more information, see the MicrosoftSQLServerSettings structure.

pub fn set_microsoft_sql_server_settings(
self,
input: Option<MicrosoftSqlServerSettings>
) -> Self

pub fn ibm_db2_settings(self, input: IbmDb2Settings) -> Self

The settings for the IBM Db2 LUW source endpoint. For more information, see the IBMDb2Settings structure.

pub fn set_ibm_db2_settings(self, input: Option<IbmDb2Settings>) -> Self

pub fn doc_db_settings(self, input: DocDbSettings) -> Self

Provides information that defines a DocumentDB endpoint.

pub fn set_doc_db_settings(self, input: Option<DocDbSettings>) -> Self

pub fn redis_settings(self, input: RedisSettings) -> Self

The settings for the Redis target endpoint. For more information, see the RedisSettings structure.

pub fn set_redis_settings(self, input: Option<RedisSettings>) -> Self

pub fn build(self) -> Endpoint

Consumes the builder and constructs a Endpoint