Struct aws_sdk_databasemigration::model::oracle_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for OracleSettings

Implementations

impl Builder

pub fn add_supplemental_logging(self, input: bool) -> Self

Set this attribute to set up table-level supplemental logging for the Oracle database. This attribute enables PRIMARY KEY supplemental logging on all tables selected for a migration task.

If you use this option, you still need to enable database-level supplemental logging.

pub fn set_add_supplemental_logging(self, input: Option<bool>) -> Self

pub fn archived_log_dest_id(self, input: i32) -> Self

Specifies the ID of the destination for the archived redo logs. This value should be the same as a number in the dest_id column of the v$archived_log view. If you work with an additional redo log destination, use the AdditionalArchivedLogDestId option to specify the additional destination ID. Doing this improves performance by ensuring that the correct logs are accessed from the outset.

pub fn set_archived_log_dest_id(self, input: Option<i32>) -> Self

pub fn additional_archived_log_dest_id(self, input: i32) -> Self

Set this attribute with ArchivedLogDestId in a primary/ standby setup. This attribute is useful in the case of a switchover. In this case, DMS needs to know which destination to get archive redo logs from to read changes. This need arises because the previous primary instance is now a standby instance after switchover.

Although DMS supports the use of the Oracle RESETLOGS option to open the database, never use RESETLOGS unless necessary. For additional information about RESETLOGS, see RMAN Data Repair Concepts in the Oracle Database Backup and Recovery User's Guide.

pub fn set_additional_archived_log_dest_id(self, input: Option<i32>) -> Self

pub fn extra_archived_log_dest_ids(self, input: impl Into<i32>) -> Self

pub fn set_extra_archived_log_dest_ids(self, input: Option<Vec<i32>>) -> Self

pub fn allow_select_nested_tables(self, input: bool) -> Self

Set this attribute to true to enable replication of Oracle tables containing columns that are nested tables or defined types.

pub fn set_allow_select_nested_tables(self, input: Option<bool>) -> Self

pub fn parallel_asm_read_threads(self, input: i32) -> Self

Set this attribute to change the number of threads that DMS configures to perform a change data capture (CDC) load using Oracle Automatic Storage Management (ASM). You can specify an integer value between 2 (the default) and 8 (the maximum). Use this attribute together with the readAheadBlocks attribute.

pub fn set_parallel_asm_read_threads(self, input: Option<i32>) -> Self

pub fn read_ahead_blocks(self, input: i32) -> Self

Set this attribute to change the number of read-ahead blocks that DMS configures to perform a change data capture (CDC) load using Oracle Automatic Storage Management (ASM). You can specify an integer value between 1000 (the default) and 200,000 (the maximum).

pub fn set_read_ahead_blocks(self, input: Option<i32>) -> Self

pub fn access_alternate_directly(self, input: bool) -> Self

Set this attribute to false in order to use the Binary Reader to capture change data for an Amazon RDS for Oracle as the source. This tells the DMS instance to not access redo logs through any specified path prefix replacement using direct file access.

pub fn set_access_alternate_directly(self, input: Option<bool>) -> Self

pub fn use_alternate_folder_for_online(self, input: bool) -> Self

Set this attribute to true in order to use the Binary Reader to capture change data for an Amazon RDS for Oracle as the source. This tells the DMS instance to use any specified prefix replacement to access all online redo logs.

pub fn set_use_alternate_folder_for_online(self, input: Option<bool>) -> Self

pub fn oracle_path_prefix(self, input: impl Into<String>) -> Self

Set this string attribute to the required value in order to use the Binary Reader to capture change data for an Amazon RDS for Oracle as the source. This value specifies the default Oracle root used to access the redo logs.

pub fn set_oracle_path_prefix(self, input: Option<String>) -> Self

pub fn use_path_prefix(self, input: impl Into<String>) -> Self

Set this string attribute to the required value in order to use the Binary Reader to capture change data for an Amazon RDS for Oracle as the source. This value specifies the path prefix used to replace the default Oracle root to access the redo logs.

pub fn set_use_path_prefix(self, input: Option<String>) -> Self

pub fn replace_path_prefix(self, input: bool) -> Self

Set this attribute to true in order to use the Binary Reader to capture change data for an Amazon RDS for Oracle as the source. This setting tells DMS instance to replace the default Oracle root with the specified usePathPrefix setting to access the redo logs.

pub fn set_replace_path_prefix(self, input: Option<bool>) -> Self

pub fn enable_homogenous_tablespace(self, input: bool) -> Self

Set this attribute to enable homogenous tablespace replication and create existing tables or indexes under the same tablespace on the target.

pub fn set_enable_homogenous_tablespace(self, input: Option<bool>) -> Self

pub fn direct_path_no_log(self, input: bool) -> Self

When set to true, this attribute helps to increase the commit rate on the Oracle target database by writing directly to tables and not writing a trail to database logs.

pub fn set_direct_path_no_log(self, input: Option<bool>) -> Self

pub fn archived_logs_only(self, input: bool) -> Self

When this field is set to Y, DMS only accesses the archived redo logs. If the archived redo logs are stored on Oracle ASM only, the DMS user account needs to be granted ASM privileges.

pub fn set_archived_logs_only(self, input: Option<bool>) -> Self

pub fn asm_password(self, input: impl Into<String>) -> Self

For an Oracle source endpoint, your Oracle Automatic Storage Management (ASM) password. You can set this value from the asm_user_password value. You set this value as part of the comma-separated value that you set to the Password request parameter when you create the endpoint to access transaction logs using Binary Reader. For more information, see Configuration for change data capture (CDC) on an Oracle source database.

pub fn set_asm_password(self, input: Option<String>) -> Self

pub fn asm_server(self, input: impl Into<String>) -> Self

For an Oracle source endpoint, your ASM server address. You can set this value from the asm_server value. You set asm_server as part of the extra connection attribute string to access an Oracle server with Binary Reader that uses ASM. For more information, see Configuration for change data capture (CDC) on an Oracle source database.

pub fn set_asm_server(self, input: Option<String>) -> Self

pub fn asm_user(self, input: impl Into<String>) -> Self

For an Oracle source endpoint, your ASM user name. You can set this value from the asm_user value. You set asm_user as part of the extra connection attribute string to access an Oracle server with Binary Reader that uses ASM. For more information, see Configuration for change data capture (CDC) on an Oracle source database.

pub fn set_asm_user(self, input: Option<String>) -> Self

pub fn char_length_semantics(self, input: CharLengthSemantics) -> Self

Specifies whether the length of a character column is in bytes or in characters. To indicate that the character column length is in characters, set this attribute to CHAR. Otherwise, the character column length is in bytes.

Example: charLengthSemantics=CHAR;

pub fn set_char_length_semantics(
self,
input: Option<CharLengthSemantics>
) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

Database name for the endpoint.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn direct_path_parallel_load(self, input: bool) -> Self

When set to true, this attribute specifies a parallel load when useDirectPathFullLoad is set to Y. This attribute also only applies when you use the DMS parallel load feature. Note that the target table cannot have any constraints or indexes.

pub fn set_direct_path_parallel_load(self, input: Option<bool>) -> Self

pub fn fail_tasks_on_lob_truncation(self, input: bool) -> Self

When set to true, this attribute causes a task to fail if the actual size of an LOB column is greater than the specified LobMaxSize.

If a task is set to limited LOB mode and this option is set to true, the task fails instead of truncating the LOB data.

pub fn set_fail_tasks_on_lob_truncation(self, input: Option<bool>) -> Self

pub fn number_datatype_scale(self, input: i32) -> Self

Specifies the number scale. You can select a scale up to 38, or you can select FLOAT. By default, the NUMBER data type is converted to precision 38, scale 10.

Example: numberDataTypeScale=12

pub fn set_number_datatype_scale(self, input: Option<i32>) -> Self

pub fn password(self, input: impl Into<String>) -> Self

Endpoint connection password.

pub fn set_password(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

Endpoint TCP port.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn read_table_space_name(self, input: bool) -> Self

When set to true, this attribute supports tablespace replication.

pub fn set_read_table_space_name(self, input: Option<bool>) -> Self

pub fn retry_interval(self, input: i32) -> Self

Specifies the number of seconds that the system waits before resending a query.

Example: retryInterval=6;

pub fn set_retry_interval(self, input: Option<i32>) -> Self

pub fn security_db_encryption(self, input: impl Into<String>) -> Self

For an Oracle source endpoint, the transparent data encryption (TDE) password required by AWM DMS to access Oracle redo logs encrypted by TDE using Binary Reader. It is also the TDE_Password part of the comma-separated value you set to the Password request parameter when you create the endpoint. The SecurityDbEncryptian setting is related to this SecurityDbEncryptionName setting. For more information, see Supported encryption methods for using Oracle as a source for DMS in the Database Migration Service User Guide.

pub fn set_security_db_encryption(self, input: Option<String>) -> Self

pub fn security_db_encryption_name(self, input: impl Into<String>) -> Self

For an Oracle source endpoint, the name of a key used for the transparent data encryption (TDE) of the columns and tablespaces in an Oracle source database that is encrypted using TDE. The key value is the value of the SecurityDbEncryption setting. For more information on setting the key name value of SecurityDbEncryptionName, see the information and example for setting the securityDbEncryptionName extra connection attribute in Supported encryption methods for using Oracle as a source for DMS in the Database Migration Service User Guide.

pub fn set_security_db_encryption_name(self, input: Option<String>) -> Self

pub fn server_name(self, input: impl Into<String>) -> Self

Fully qualified domain name of the endpoint.

pub fn set_server_name(self, input: Option<String>) -> Self

pub fn spatial_data_option_to_geo_json_function_name(
self,
input: impl Into<String>
) -> Self

Use this attribute to convert SDO_GEOMETRY to GEOJSON format. By default, DMS calls the SDO2GEOJSON custom function if present and accessible. Or you can create your own custom function that mimics the operation of SDOGEOJSON and set SpatialDataOptionToGeoJsonFunctionName to call it instead.

pub fn set_spatial_data_option_to_geo_json_function_name(
self,
input: Option<String>
) -> Self

pub fn standby_delay_time(self, input: i32) -> Self

Use this attribute to specify a time in minutes for the delay in standby sync. If the source is an Oracle Active Data Guard standby database, use this attribute to specify the time lag between primary and standby databases.

In DMS, you can create an Oracle CDC task that uses an Active Data Guard standby instance as a source for replicating ongoing changes. Doing this eliminates the need to connect to an active database that might be in production.

pub fn set_standby_delay_time(self, input: Option<i32>) -> Self

pub fn username(self, input: impl Into<String>) -> Self

Endpoint connection user name.

pub fn set_username(self, input: Option<String>) -> Self

pub fn use_b_file(self, input: bool) -> Self

Set this attribute to Y to capture change data using the Binary Reader utility. Set UseLogminerReader to N to set this attribute to Y. To use Binary Reader with Amazon RDS for Oracle as the source, you set additional attributes. For more information about using this setting with Oracle Automatic Storage Management (ASM), see Using Oracle LogMiner or DMS Binary Reader for CDC.

pub fn set_use_b_file(self, input: Option<bool>) -> Self

pub fn use_direct_path_full_load(self, input: bool) -> Self

Set this attribute to Y to have DMS use a direct path full load. Specify this value to use the direct path protocol in the Oracle Call Interface (OCI). By using this OCI protocol, you can bulk-load Oracle target tables during a full load.

pub fn set_use_direct_path_full_load(self, input: Option<bool>) -> Self

pub fn use_logminer_reader(self, input: bool) -> Self

Set this attribute to Y to capture change data using the Oracle LogMiner utility (the default). Set this attribute to N if you want to access the redo logs as a binary file. When you set UseLogminerReader to N, also set UseBfile to Y. For more information on this setting and using Oracle ASM, see Using Oracle LogMiner or DMS Binary Reader for CDC in the DMS User Guide.

pub fn set_use_logminer_reader(self, input: Option<bool>) -> Self

pub fn secrets_manager_access_role_arn(self, input: impl Into<String>) -> Self

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the Oracle endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

pub fn set_secrets_manager_access_role_arn(self, input: Option<String>) -> Self

pub fn secrets_manager_secret_id(self, input: impl Into<String>) -> Self

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the Oracle endpoint connection details.

pub fn set_secrets_manager_secret_id(self, input: Option<String>) -> Self

pub fn secrets_manager_oracle_asm_access_role_arn(
self,
input: impl Into<String>
) -> Self

Required only if your Oracle endpoint uses Advanced Storage Manager (ASM). The full ARN of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the SecretsManagerOracleAsmSecret. This SecretsManagerOracleAsmSecret has the secret value that allows access to the Oracle ASM of the endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerOracleAsmSecretId. Or you can specify clear-text values for AsmUserName, AsmPassword, and AsmServerName. You can't specify both. For more information on creating this SecretsManagerOracleAsmSecret and the SecretsManagerOracleAsmAccessRoleArn and SecretsManagerOracleAsmSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

pub fn set_secrets_manager_oracle_asm_access_role_arn(
self,
input: Option<String>
) -> Self

pub fn secrets_manager_oracle_asm_secret_id(
self,
input: impl Into<String>
) -> Self

Required only if your Oracle endpoint uses Advanced Storage Manager (ASM). The full ARN, partial ARN, or friendly name of the SecretsManagerOracleAsmSecret that contains the Oracle ASM connection details for the Oracle endpoint.

pub fn set_secrets_manager_oracle_asm_secret_id(
self,
input: Option<String>
) -> Self

pub fn build(self) -> OracleSettings

Consumes the builder and constructs a OracleSettings