Struct aws_sdk_databasemigration::model::postgre_sql_settings::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for PostgreSqlSettings

Implementations

impl Builder

pub fn after_connect_script(self, input: impl Into<String>) -> Self

For use with change data capture (CDC) only, this attribute has DMS bypass foreign keys and user triggers to reduce the time it takes to bulk load data.

Example: afterConnectScript=SET session_replication_role='replica'

pub fn set_after_connect_script(self, input: Option<String>) -> Self

pub fn capture_ddls(self, input: bool) -> Self

To capture DDL events, DMS creates various artifacts in the PostgreSQL database when the task starts. You can later remove these artifacts.

If this value is set to N, you don't have to create tables or triggers on the source database.

pub fn set_capture_ddls(self, input: Option<bool>) -> Self

pub fn max_file_size(self, input: i32) -> Self

Specifies the maximum size (in KB) of any .csv file used to transfer data to PostgreSQL.

Example: maxFileSize=512

pub fn set_max_file_size(self, input: Option<i32>) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

Database name for the endpoint.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn ddl_artifacts_schema(self, input: impl Into<String>) -> Self

The schema in which the operational DDL database artifacts are created.

Example: ddlArtifactsSchema=xyzddlschema;

pub fn set_ddl_artifacts_schema(self, input: Option<String>) -> Self

pub fn execute_timeout(self, input: i32) -> Self

Sets the client statement timeout for the PostgreSQL instance, in seconds. The default value is 60 seconds.

Example: executeTimeout=100;

pub fn set_execute_timeout(self, input: Option<i32>) -> Self

pub fn fail_tasks_on_lob_truncation(self, input: bool) -> Self

When set to true, this value causes a task to fail if the actual size of a LOB column is greater than the specified LobMaxSize.

If task is set to Limited LOB mode and this option is set to true, the task fails instead of truncating the LOB data.

pub fn set_fail_tasks_on_lob_truncation(self, input: Option<bool>) -> Self

pub fn heartbeat_enable(self, input: bool) -> Self

The write-ahead log (WAL) heartbeat feature mimics a dummy transaction. By doing this, it prevents idle logical replication slots from holding onto old WAL logs, which can result in storage full situations on the source. This heartbeat keeps restart_lsn moving and prevents storage full scenarios.

pub fn set_heartbeat_enable(self, input: Option<bool>) -> Self

pub fn heartbeat_schema(self, input: impl Into<String>) -> Self

Sets the schema in which the heartbeat artifacts are created.

pub fn set_heartbeat_schema(self, input: Option<String>) -> Self

pub fn heartbeat_frequency(self, input: i32) -> Self

Sets the WAL heartbeat frequency (in minutes).

pub fn set_heartbeat_frequency(self, input: Option<i32>) -> Self

pub fn password(self, input: impl Into<String>) -> Self

Endpoint connection password.

pub fn set_password(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

Endpoint TCP port.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn server_name(self, input: impl Into<String>) -> Self

Fully qualified domain name of the endpoint.

pub fn set_server_name(self, input: Option<String>) -> Self

pub fn username(self, input: impl Into<String>) -> Self

Endpoint connection user name.

pub fn set_username(self, input: Option<String>) -> Self

pub fn slot_name(self, input: impl Into<String>) -> Self

Sets the name of a previously created logical replication slot for a change data capture (CDC) load of the PostgreSQL source instance.

When used with the CdcStartPosition request parameter for the DMS API , this attribute also makes it possible to use native CDC start points. DMS verifies that the specified logical replication slot exists before starting the CDC load task. It also verifies that the task was created with a valid setting of CdcStartPosition. If the specified slot doesn't exist or the task doesn't have a valid CdcStartPosition setting, DMS raises an error.

For more information about setting the CdcStartPosition request parameter, see Determining a CDC native start point in the Database Migration Service User Guide. For more information about using CdcStartPosition, see CreateReplicationTask, StartReplicationTask, and ModifyReplicationTask.

pub fn set_slot_name(self, input: Option<String>) -> Self

pub fn plugin_name(self, input: PluginNameValue) -> Self

Specifies the plugin to use to create a replication slot.

pub fn set_plugin_name(self, input: Option<PluginNameValue>) -> Self

pub fn secrets_manager_access_role_arn(self, input: impl Into<String>) -> Self

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the PostgreSQL endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

pub fn set_secrets_manager_access_role_arn(self, input: Option<String>) -> Self

pub fn secrets_manager_secret_id(self, input: impl Into<String>) -> Self

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the PostgreSQL endpoint connection details.

pub fn set_secrets_manager_secret_id(self, input: Option<String>) -> Self

pub fn build(self) -> PostgreSqlSettings

Consumes the builder and constructs a PostgreSqlSettings