Struct aws_sdk_databasemigration::model::replication_task::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ReplicationTask

Implementations

impl Builder

pub fn replication_task_identifier(self, input: impl Into<String>) -> Self

The user-assigned replication task identifier or name.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

pub fn set_replication_task_identifier(self, input: Option<String>) -> Self

pub fn source_endpoint_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) that uniquely identifies the endpoint.

pub fn set_source_endpoint_arn(self, input: Option<String>) -> Self

pub fn target_endpoint_arn(self, input: impl Into<String>) -> Self

The ARN that uniquely identifies the endpoint.

pub fn set_target_endpoint_arn(self, input: Option<String>) -> Self

pub fn replication_instance_arn(self, input: impl Into<String>) -> Self

The ARN of the replication instance.

pub fn set_replication_instance_arn(self, input: Option<String>) -> Self

pub fn migration_type(self, input: MigrationTypeValue) -> Self

The type of migration.

pub fn set_migration_type(self, input: Option<MigrationTypeValue>) -> Self

pub fn table_mappings(self, input: impl Into<String>) -> Self

Table mappings specified in the task.

pub fn set_table_mappings(self, input: Option<String>) -> Self

pub fn replication_task_settings(self, input: impl Into<String>) -> Self

The settings for the replication task.

pub fn set_replication_task_settings(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The status of the replication task. This response parameter can return one of the following values:

"moving" – The task is being moved in response to running the MoveReplicationTask operation.

"creating" – The task is being created in response to running the CreateReplicationTask operation.

"deleting" – The task is being deleted in response to running the DeleteReplicationTask operation.

"failed" – The task failed to successfully complete the database migration in response to running the StartReplicationTask operation.

"failed-move" – The task failed to move in response to running the MoveReplicationTask operation.

"modifying" – The task definition is being modified in response to running the ModifyReplicationTask operation.

"ready" – The task is in a ready state where it can respond to other task operations, such as StartReplicationTask or DeleteReplicationTask .

"running" – The task is performing a database migration in response to running the StartReplicationTask operation.

"starting" – The task is preparing to perform a database migration in response to running the StartReplicationTask operation.

"stopped" – The task has stopped in response to running the StopReplicationTask operation.

"stopping" – The task is preparing to stop in response to running the StopReplicationTask operation.

"testing" – The database migration specified for this task is being tested in response to running either the StartReplicationTaskAssessmentRun or the StartReplicationTaskAssessment operation.

StartReplicationTaskAssessmentRun is an improved premigration task assessment operation. The StartReplicationTaskAssessment operation assesses data type compatibility only between the source and target database of a given migration task. In contrast, StartReplicationTaskAssessmentRun enables you to specify a variety of premigration task assessments in addition to data type compatibility. These assessments include ones for the validity of primary key definitions and likely issues with database migration performance, among others.

pub fn set_status(self, input: Option<String>) -> Self

pub fn last_failure_message(self, input: impl Into<String>) -> Self

The last error (failure) message generated for the replication task.

pub fn set_last_failure_message(self, input: Option<String>) -> Self

pub fn stop_reason(self, input: impl Into<String>) -> Self

The reason the replication task was stopped. This response parameter can return one of the following values:

"STOP_REASON_FULL_LOAD_COMPLETED" – Full-load migration completed.

"STOP_REASON_CACHED_CHANGES_APPLIED" – Change data capture (CDC) load completed.

"STOP_REASON_CACHED_CHANGES_NOT_APPLIED" – In a full-load and CDC migration, the full load stopped as specified before starting the CDC migration.

"STOP_REASON_SERVER_TIME" – The migration stopped at the specified server time.

pub fn set_stop_reason(self, input: Option<String>) -> Self

pub fn replication_task_creation_date(self, input: Instant) -> Self

The date the replication task was created.

pub fn set_replication_task_creation_date(self, input: Option<Instant>) -> Self

pub fn replication_task_start_date(self, input: Instant) -> Self

The date the replication task is scheduled to start.

pub fn set_replication_task_start_date(self, input: Option<Instant>) -> Self

pub fn cdc_start_position(self, input: impl Into<String>) -> Self

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want the CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1#27#mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:1876#0#0#*#0#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

pub fn set_cdc_start_position(self, input: Option<String>) -> Self

pub fn cdc_stop_position(self, input: impl Into<String>) -> Self

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

pub fn set_cdc_stop_position(self, input: Option<String>) -> Self

pub fn recovery_checkpoint(self, input: impl Into<String>) -> Self

Indicates the last checkpoint that occurred during a change data capture (CDC) operation. You can provide this value to the CdcStartPosition parameter to start a CDC operation that begins at that checkpoint.

pub fn set_recovery_checkpoint(self, input: Option<String>) -> Self

pub fn replication_task_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the replication task.

pub fn set_replication_task_arn(self, input: Option<String>) -> Self

pub fn replication_task_stats(self, input: ReplicationTaskStats) -> Self

The statistics for the task, including elapsed time, tables loaded, and table errors.

pub fn set_replication_task_stats(
self,
input: Option<ReplicationTaskStats>
) -> Self

pub fn task_data(self, input: impl Into<String>) -> Self

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

pub fn set_task_data(self, input: Option<String>) -> Self

pub fn target_replication_instance_arn(self, input: impl Into<String>) -> Self

The ARN of the replication instance to which this task is moved in response to running the MoveReplicationTask operation. Otherwise, this response parameter isn't a member of the ReplicationTask object.

pub fn set_target_replication_instance_arn(self, input: Option<String>) -> Self

pub fn build(self) -> ReplicationTask

Consumes the builder and constructs a ReplicationTask