Struct aws_sdk_databasemigration::model::Endpoint

#[non_exhaustive]
pub struct Endpoint {

    pub endpoint_identifier: Option<String>,
    pub endpoint_type: Option<ReplicationEndpointTypeValue>,
    pub engine_name: Option<String>,
    pub engine_display_name: Option<String>,
    pub username: Option<String>,
    pub server_name: Option<String>,
    pub port: Option<i32>,
    pub database_name: Option<String>,
    pub extra_connection_attributes: Option<String>,
    pub status: Option<String>,
    pub kms_key_id: Option<String>,
    pub endpoint_arn: Option<String>,
    pub certificate_arn: Option<String>,
    pub ssl_mode: Option<DmsSslModeValue>,
    pub service_access_role_arn: Option<String>,
    pub external_table_definition: Option<String>,
    pub external_id: Option<String>,
    pub dynamo_db_settings: Option<DynamoDbSettings>,
    pub s3_settings: Option<S3Settings>,
    pub dms_transfer_settings: Option<DmsTransferSettings>,
    pub mongo_db_settings: Option<MongoDbSettings>,
    pub kinesis_settings: Option<KinesisSettings>,
    pub kafka_settings: Option<KafkaSettings>,
    pub elasticsearch_settings: Option<ElasticsearchSettings>,
    pub neptune_settings: Option<NeptuneSettings>,
    pub redshift_settings: Option<RedshiftSettings>,
    pub postgre_sql_settings: Option<PostgreSqlSettings>,
    pub my_sql_settings: Option<MySqlSettings>,
    pub oracle_settings: Option<OracleSettings>,
    pub sybase_settings: Option<SybaseSettings>,
    pub microsoft_sql_server_settings: Option<MicrosoftSqlServerSettings>,
    pub ibm_db2_settings: Option<IbmDb2Settings>,
    pub doc_db_settings: Option<DocDbSettings>,
    pub redis_settings: Option<RedisSettings>,

}

Describes an endpoint of a database instance in response to operations such as the following:

CreateEndpoint

DescribeEndpoint

ModifyEndpoint

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

endpoint_identifier: Option<String>

The database endpoint identifier. Identifiers must begin with a letter and must contain only ASCII letters, digits, and hyphens. They can't end with a hyphen or contain two consecutive hyphens.

endpoint_type: Option<ReplicationEndpointTypeValue>

The type of endpoint. Valid values are source and target.

engine_name: Option<String>

The database engine name. Valid values, depending on the EndpointType, include "mysql", "oracle", "postgres", "mariadb", "aurora", "aurora-postgresql", "redshift", "s3", "db2", "azuredb", "sybase", "dynamodb", "mongodb", "kinesis", "kafka", "elasticsearch", "documentdb", "sqlserver", and "neptune".

engine_display_name: Option<String>

The expanded name for the engine name. For example, if the EngineName parameter is "aurora," this value would be "Amazon Aurora MySQL."

username: Option<String>

The user name used to connect to the endpoint.

server_name: Option<String>

The name of the server at the endpoint.

port: Option<i32>

The port value used to access the endpoint.

database_name: Option<String>

The name of the database at the endpoint.

extra_connection_attributes: Option<String>

Additional connection attributes used to connect to the endpoint.

status: Option<String>

The status of the endpoint.

kms_key_id: Option<String>

An KMS key identifier that is used to encrypt the connection parameters for the endpoint.

If you don't specify a value for the KmsKeyId parameter, then DMS uses your default encryption key.

KMS creates the default encryption key for your Amazon Web Services account. Your Amazon Web Services account has a different default encryption key for each Amazon Web Services Region.

endpoint_arn: Option<String>

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

certificate_arn: Option<String>

The Amazon Resource Name (ARN) used for SSL connection to the endpoint.

ssl_mode: Option<DmsSslModeValue>

The SSL mode used to connect to the endpoint. The default value is none.

service_access_role_arn: Option<String>

The Amazon Resource Name (ARN) used by the service to access the IAM role. The role must allow the iam:PassRole action.

external_table_definition: Option<String>

The external table definition.

external_id: Option<String>

Value returned by a call to CreateEndpoint that can be used for cross-account validation. Use it on a subsequent call to CreateEndpoint to create the endpoint with a cross-account.

dynamo_db_settings: Option<DynamoDbSettings>

The settings for the DynamoDB target endpoint. For more information, see the DynamoDBSettings structure.

s3_settings: Option<S3Settings>

The settings for the S3 target endpoint. For more information, see the S3Settings structure.

dms_transfer_settings: Option<DmsTransferSettings>

The settings in JSON format for the DMS transfer type of source endpoint.

Possible settings include the following:

ServiceAccessRoleArn - The IAM role that has permission to access the Amazon S3 bucket. The role must allow the iam:PassRole action.

BucketName - The name of the S3 bucket to use.

Shorthand syntax for these settings is as follows: ServiceAccessRoleArn=string,BucketName=string,

JSON syntax for these settings is as follows: { "ServiceAccessRoleArn": "string", "BucketName": "string"}

mongo_db_settings: Option<MongoDbSettings>

The settings for the MongoDB source endpoint. For more information, see the MongoDbSettings structure.

kinesis_settings: Option<KinesisSettings>

The settings for the Amazon Kinesis target endpoint. For more information, see the KinesisSettings structure.

kafka_settings: Option<KafkaSettings>

The settings for the Apache Kafka target endpoint. For more information, see the KafkaSettings structure.

elasticsearch_settings: Option<ElasticsearchSettings>

The settings for the Elasticsearch source endpoint. For more information, see the ElasticsearchSettings structure.

neptune_settings: Option<NeptuneSettings>

The settings for the Amazon Neptune target endpoint. For more information, see the NeptuneSettings structure.

redshift_settings: Option<RedshiftSettings>

Settings for the Amazon Redshift endpoint.

postgre_sql_settings: Option<PostgreSqlSettings>

The settings for the PostgreSQL source and target endpoint. For more information, see the PostgreSQLSettings structure.

my_sql_settings: Option<MySqlSettings>

The settings for the MySQL source and target endpoint. For more information, see the MySQLSettings structure.

oracle_settings: Option<OracleSettings>

The settings for the Oracle source and target endpoint. For more information, see the OracleSettings structure.

sybase_settings: Option<SybaseSettings>

The settings for the SAP ASE source and target endpoint. For more information, see the SybaseSettings structure.

microsoft_sql_server_settings: Option<MicrosoftSqlServerSettings>

The settings for the Microsoft SQL Server source and target endpoint. For more information, see the MicrosoftSQLServerSettings structure.

ibm_db2_settings: Option<IbmDb2Settings>

The settings for the IBM Db2 LUW source endpoint. For more information, see the IBMDb2Settings structure.

doc_db_settings: Option<DocDbSettings>

Provides information that defines a DocumentDB endpoint.

redis_settings: Option<RedisSettings>

The settings for the Redis target endpoint. For more information, see the RedisSettings structure.

Implementations

impl Endpoint

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Endpoint