Struct aws_sdk_databasemigration::model::MySqlSettings

#[non_exhaustive]
pub struct MySqlSettings {

    pub after_connect_script: Option<String>,
    pub clean_source_metadata_on_mismatch: Option<bool>,
    pub database_name: Option<String>,
    pub events_poll_interval: Option<i32>,
    pub target_db_type: Option<TargetDbType>,
    pub max_file_size: Option<i32>,
    pub parallel_load_threads: Option<i32>,
    pub password: Option<String>,
    pub port: Option<i32>,
    pub server_name: Option<String>,
    pub server_timezone: Option<String>,
    pub username: Option<String>,
    pub secrets_manager_access_role_arn: Option<String>,
    pub secrets_manager_secret_id: Option<String>,

}

Provides information that defines a MySQL endpoint.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

after_connect_script: Option<String>

Specifies a script to run immediately after DMS connects to the endpoint. The migration task continues running regardless if the SQL statement succeeds or fails.

For this parameter, provide the code of the script itself, not the name of a file containing the script.

clean_source_metadata_on_mismatch: Option<bool>

Adjusts the behavior of DMS when migrating from an SQL Server source database that is hosted as part of an Always On availability group cluster. If you need DMS to poll all the nodes in the Always On cluster for transaction backups, set this attribute to false.

database_name: Option<String>

Database name for the endpoint. For a MySQL source or target endpoint, don't explicitly specify the database using the DatabaseName request parameter on either the CreateEndpoint or ModifyEndpoint API call. Specifying DatabaseName when you create or modify a MySQL endpoint replicates all the task tables to this single database. For MySQL endpoints, you specify the database only when you specify the schema in the table-mapping rules of the DMS task.

events_poll_interval: Option<i32>

Specifies how often to check the binary log for new changes/events when the database is idle.

Example: eventsPollInterval=5;

In the example, DMS checks for changes in the binary logs every five seconds.

target_db_type: Option<TargetDbType>

Specifies where to migrate source tables on the target, either to a single database or multiple databases.

Example: targetDbType=MULTIPLE_DATABASES

max_file_size: Option<i32>

Specifies the maximum size (in KB) of any .csv file used to transfer data to a MySQL-compatible database.

Example: maxFileSize=512

parallel_load_threads: Option<i32>

Improves performance when loading data into the MySQL-compatible target database. Specifies how many threads to use to load the data into the MySQL-compatible target database. Setting a large number of threads can have an adverse effect on database performance, because a separate connection is required for each thread.

Example: parallelLoadThreads=1

password: Option<String>

Endpoint connection password.

port: Option<i32>

Endpoint TCP port.

server_name: Option<String>

Fully qualified domain name of the endpoint.

server_timezone: Option<String>

Specifies the time zone for the source MySQL database.

Example: serverTimezone=US/Pacific;

Note: Do not enclose time zones in single quotes.

username: Option<String>

Endpoint connection user name.

secrets_manager_access_role_arn: Option<String>

The full Amazon Resource Name (ARN) of the IAM role that specifies DMS as the trusted entity and grants the required permissions to access the value in SecretsManagerSecret. The role must allow the iam:PassRole action. SecretsManagerSecret has the value of the Amazon Web Services Secrets Manager secret that allows access to the MySQL endpoint.

You can specify one of two sets of values for these permissions. You can specify the values for this setting and SecretsManagerSecretId. Or you can specify clear-text values for UserName, Password, ServerName, and Port. You can't specify both. For more information on creating this SecretsManagerSecret and the SecretsManagerAccessRoleArn and SecretsManagerSecretId required to access it, see Using secrets to access Database Migration Service resources in the Database Migration Service User Guide.

secrets_manager_secret_id: Option<String>

The full ARN, partial ARN, or friendly name of the SecretsManagerSecret that contains the MySQL endpoint connection details.

Implementations

impl MySqlSettings

pub fn builder() -> Builder

Creates a new builder-style object to manufacture MySqlSettings