Struct aws_sdk_databasemigration::model::ReplicationTask

#[non_exhaustive]
pub struct ReplicationTask {

    pub replication_task_identifier: Option<String>,
    pub source_endpoint_arn: Option<String>,
    pub target_endpoint_arn: Option<String>,
    pub replication_instance_arn: Option<String>,
    pub migration_type: Option<MigrationTypeValue>,
    pub table_mappings: Option<String>,
    pub replication_task_settings: Option<String>,
    pub status: Option<String>,
    pub last_failure_message: Option<String>,
    pub stop_reason: Option<String>,
    pub replication_task_creation_date: Option<Instant>,
    pub replication_task_start_date: Option<Instant>,
    pub cdc_start_position: Option<String>,
    pub cdc_stop_position: Option<String>,
    pub recovery_checkpoint: Option<String>,
    pub replication_task_arn: Option<String>,
    pub replication_task_stats: Option<ReplicationTaskStats>,
    pub task_data: Option<String>,
    pub target_replication_instance_arn: Option<String>,

}

Provides information that describes a replication task created by the CreateReplicationTask operation.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

replication_task_identifier: Option<String>

The user-assigned replication task identifier or name.

Constraints:

Must contain 1-255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

source_endpoint_arn: Option<String>

The Amazon Resource Name (ARN) that uniquely identifies the endpoint.

target_endpoint_arn: Option<String>

The ARN that uniquely identifies the endpoint.

replication_instance_arn: Option<String>

The ARN of the replication instance.

migration_type: Option<MigrationTypeValue>

The type of migration.

table_mappings: Option<String>

Table mappings specified in the task.

replication_task_settings: Option<String>

The settings for the replication task.

status: Option<String>

The status of the replication task. This response parameter can return one of the following values:

"moving" – The task is being moved in response to running the MoveReplicationTask operation.

"creating" – The task is being created in response to running the CreateReplicationTask operation.

"deleting" – The task is being deleted in response to running the DeleteReplicationTask operation.

"failed" – The task failed to successfully complete the database migration in response to running the StartReplicationTask operation.

"failed-move" – The task failed to move in response to running the MoveReplicationTask operation.

"modifying" – The task definition is being modified in response to running the ModifyReplicationTask operation.

"ready" – The task is in a ready state where it can respond to other task operations, such as StartReplicationTask or DeleteReplicationTask .

"running" – The task is performing a database migration in response to running the StartReplicationTask operation.

"starting" – The task is preparing to perform a database migration in response to running the StartReplicationTask operation.

"stopped" – The task has stopped in response to running the StopReplicationTask operation.

"stopping" – The task is preparing to stop in response to running the StopReplicationTask operation.

"testing" – The database migration specified for this task is being tested in response to running either the StartReplicationTaskAssessmentRun or the StartReplicationTaskAssessment operation.

StartReplicationTaskAssessmentRun is an improved premigration task assessment operation. The StartReplicationTaskAssessment operation assesses data type compatibility only between the source and target database of a given migration task. In contrast, StartReplicationTaskAssessmentRun enables you to specify a variety of premigration task assessments in addition to data type compatibility. These assessments include ones for the validity of primary key definitions and likely issues with database migration performance, among others.

last_failure_message: Option<String>

The last error (failure) message generated for the replication task.

stop_reason: Option<String>

The reason the replication task was stopped. This response parameter can return one of the following values:

"STOP_REASON_FULL_LOAD_COMPLETED" – Full-load migration completed.

"STOP_REASON_CACHED_CHANGES_APPLIED" – Change data capture (CDC) load completed.

"STOP_REASON_CACHED_CHANGES_NOT_APPLIED" – In a full-load and CDC migration, the full load stopped as specified before starting the CDC migration.

"STOP_REASON_SERVER_TIME" – The migration stopped at the specified server time.

replication_task_creation_date: Option<Instant>

The date the replication task was created.

replication_task_start_date: Option<Instant>

The date the replication task is scheduled to start.

cdc_start_position: Option<String>

Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want the CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position "checkpoint:V1#27#mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:1876#0#0#*#0#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

cdc_stop_position: Option<String>

Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position “server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time: 2018-02-09T12:12:12 “

recovery_checkpoint: Option<String>

Indicates the last checkpoint that occurred during a change data capture (CDC) operation. You can provide this value to the CdcStartPosition parameter to start a CDC operation that begins at that checkpoint.

replication_task_arn: Option<String>

The Amazon Resource Name (ARN) of the replication task.

replication_task_stats: Option<ReplicationTaskStats>

The statistics for the task, including elapsed time, tables loaded, and table errors.

task_data: Option<String>

Supplemental information that the task requires to migrate the data for certain source and target endpoints. For more information, see Specifying Supplemental Data for Task Settings in the Database Migration Service User Guide.

target_replication_instance_arn: Option<String>

The ARN of the replication instance to which this task is moved in response to running the MoveReplicationTask operation. Otherwise, this response parameter isn't a member of the ReplicationTask object.

Implementations

impl ReplicationTask

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ReplicationTask