Module aws_sdk_databasemigration::operation

Structs

AddTagsToResource

Adds metadata tags to an DMS resource, including replication instance, endpoint,

ApplyPendingMaintenanceAction

Applies a pending maintenance action to a resource (for example, to a replication instance).

CancelReplicationTaskAssessmentRun

Cancels a single premigration assessment run.

CreateEndpoint

Creates an endpoint using the provided settings.

CreateEventSubscription

Creates an DMS event notification subscription.

CreateReplicationInstance

Creates the replication instance using the specified parameters.

CreateReplicationSubnetGroup

Creates a replication subnet group given a list of the subnet IDs in a VPC.

CreateReplicationTask

Creates a replication task using the specified parameters.

DeleteCertificate

Deletes the specified certificate.

DeleteConnection

Deletes the connection between a replication instance and an endpoint.

DeleteEndpoint

Deletes the specified endpoint.

DeleteEventSubscription

Deletes an DMS event subscription.

DeleteReplicationInstance

Deletes the specified replication instance.

DeleteReplicationSubnetGroup

Deletes a subnet group.

DeleteReplicationTask

Deletes the specified replication task.

DeleteReplicationTaskAssessmentRun

Deletes the record of a single premigration assessment run.

DescribeAccountAttributes

Lists all of the DMS attributes for a customer account. These attributes include DMS

DescribeApplicableIndividualAssessments

Provides a list of individual assessments that you can specify for a new premigration

DescribeCertificates

Provides a description of the certificate.

DescribeConnections

Describes the status of the connections that have been made between the replication

DescribeEndpointSettings

Returns information about the possible endpoint settings available

DescribeEndpointTypes

Returns information about the type of endpoints available.

DescribeEndpoints

Returns information about the endpoints for your account in the current region.

DescribeEventCategories

Lists categories for all event source types, or, if specified, for a specified source

DescribeEventSubscriptions

Lists all the event subscriptions for a customer account. The description of a

DescribeEvents

Lists events for a given source identifier and source type. You can also specify a

DescribeOrderableReplicationInstances

Returns information about the replication instance types that can be created in the

DescribePendingMaintenanceActions

For internal use only

DescribeRefreshSchemasStatus

Returns the status of the RefreshSchemas operation.

DescribeReplicationInstanceTaskLogs

Returns information about the task logs for the specified task.

DescribeReplicationInstances

Returns information about replication instances for your account in the current

DescribeReplicationSubnetGroups

Returns information about the replication subnet groups.

DescribeReplicationTaskAssessmentResults

Returns the task assessment results from the Amazon S3 bucket that DMS creates in your

DescribeReplicationTaskAssessmentRuns

Returns a paginated list of premigration assessment runs based on filter

DescribeReplicationTaskIndividualAssessments

Returns a paginated list of individual assessments based on filter settings.

DescribeReplicationTasks

Returns information about replication tasks for your account in the current

DescribeSchemas

Returns information about the schema for the specified endpoint.

DescribeTableStatistics

Returns table statistics on the database migration task, including table name, rows

ImportCertificate

Uploads the specified certificate.

ListTagsForResource

Lists all metadata tags attached to an DMS resource, including

ModifyEndpoint

Modifies the specified endpoint.

ModifyEventSubscription

Modifies an existing DMS event notification subscription.

ModifyReplicationInstance

Modifies the replication instance to apply new settings. You can change one or more

ModifyReplicationSubnetGroup

Modifies the settings for the specified replication subnet group.

ModifyReplicationTask

Modifies the specified replication task.

MoveReplicationTask

Moves a replication task from its current replication instance to a different target

RebootReplicationInstance

Reboots a replication instance. Rebooting results in a momentary outage, until the

RefreshSchemas

Populates the schema for the specified endpoint. This is an asynchronous operation and

ReloadTables

Reloads the target database table with the source data.

RemoveTagsFromResource

Removes metadata tags from an DMS resource, including replication instance,

StartReplicationTask

Starts the replication task.

StartReplicationTaskAssessment

Starts the replication task assessment for unsupported data types in the source

StartReplicationTaskAssessmentRun

Starts a new premigration assessment run for one or more individual assessments

StopReplicationTask

Stops the replication task.

TestConnection

Tests the connection between the replication instance and the endpoint.